requirements.txt
python-telegram-bot==13.15
requests
python-dotenv
urllib3<2.0